# project/server/docs/swagger/htmlSwag/swagger.json
{
  "basePath" : "/",
  "definitions" : {
...
    "DataAuth" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "string"
        },
        "ico" : {
          "format" : "int32",
          "type" : "integer"
        },
        "id" : {
          "type" : "string"
        },
        "idL" : {
          "format" : "int64",
          "type" : "integer"
        },
        "login" : {
          "type" : "string"
        },
        "mod" : {
          "type" : "string"
        },
        "notifToken" : {
          "type" : "string"
        },
        "par" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "permis" : {
          "type" : "boolean"
        },
        "secFr" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "uuid" : {
          "type" : "string"
        }
      }
    },
...
    "DataParents" : {
      "type" : "object",
      "properties" : {
        "bod" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "group" : {
          "format" : "int64",
          "type" : "integer"
        },
        "id" : {
          "format" : "int64",
          "type" : "integer"
        },
        "name" : {
          "type" : "string"
        },
        "uuid" : {
          "type" : "string"
        }
      }
    },
...
    "DataTeachers" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "format" : "int64",
          "type" : "integer"
        },
        "name" : {
          "type" : "string"
        },
        "uuid" : {
          "type" : "string"
        }
      }
    },
...
    "JsonArray" : {
      "type" : "object",
      "properties" : {
        "asBigDecimal" : {
          "type" : "number"
        },
        "asBigInteger" : {
          "type" : "integer"
        },
        "asBoolean" : {
          "type" : "boolean"
        },
        "asByte" : {
          "format" : "byte",
          "type" : "string"
        },
        "asCharacter" : {
          "type" : "string"
        },
        "asDouble" : {
          "format" : "double",
          "type" : "number"
        },
        "asFloat" : {
          "format" : "float",
          "type" : "number"
        },
        "asInt" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asJsonArray" : {
          "$ref" : "#/definitions/JsonArray"
        },
        "asJsonNull" : {
          "$ref" : "#/definitions/JsonNull"
        },
        "asJsonObject" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "asJsonPrimitive" : {
          "$ref" : "#/definitions/JsonPrimitive"
        },
        "asLong" : {
          "format" : "int64",
          "type" : "integer"
        },
        "asNumber" : {
          "$ref" : "#/definitions/Number"
        },
        "asShort" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asString" : {
          "type" : "string"
        },
        "empty" : {
          "type" : "boolean"
        },
        "jsonArray" : {
          "type" : "boolean"
        },
        "jsonNull" : {
          "type" : "boolean"
        },
        "jsonObject" : {
          "type" : "boolean"
        },
        "jsonPrimitive" : {
          "type" : "boolean"
        }
      }
    },
    "JsonNull" : {
      "type" : "object",
      "properties" : {
        "asBigDecimal" : {
          "type" : "number"
        },
        "asBigInteger" : {
          "type" : "integer"
        },
        "asBoolean" : {
          "type" : "boolean"
        },
        "asByte" : {
          "format" : "byte",
          "type" : "string"
        },
        "asCharacter" : {
          "type" : "string"
        },
        "asDouble" : {
          "format" : "double",
          "type" : "number"
        },
        "asFloat" : {
          "format" : "float",
          "type" : "number"
        },
        "asInt" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asJsonArray" : {
          "$ref" : "#/definitions/JsonArray"
        },
        "asJsonNull" : {
          "$ref" : "#/definitions/JsonNull"
        },
        "asJsonObject" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "asJsonPrimitive" : {
          "$ref" : "#/definitions/JsonPrimitive"
        },
        "asLong" : {
          "format" : "int64",
          "type" : "integer"
        },
        "asNumber" : {
          "$ref" : "#/definitions/Number"
        },
        "asShort" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asString" : {
          "type" : "string"
        },
        "jsonArray" : {
          "type" : "boolean"
        },
        "jsonNull" : {
          "type" : "boolean"
        },
        "jsonObject" : {
          "type" : "boolean"
        },
        "jsonPrimitive" : {
          "type" : "boolean"
        }
      }
    },
    "JsonObject" : {
      "type" : "object",
      "properties" : {
        "asBigDecimal" : {
          "type" : "number"
        },
        "asBigInteger" : {
          "type" : "integer"
        },
        "asBoolean" : {
          "type" : "boolean"
        },
        "asByte" : {
          "format" : "byte",
          "type" : "string"
        },
        "asCharacter" : {
          "type" : "string"
        },
        "asDouble" : {
          "format" : "double",
          "type" : "number"
        },
        "asFloat" : {
          "format" : "float",
          "type" : "number"
        },
        "asInt" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asJsonArray" : {
          "$ref" : "#/definitions/JsonArray"
        },
        "asJsonNull" : {
          "$ref" : "#/definitions/JsonNull"
        },
        "asJsonObject" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "asJsonPrimitive" : {
          "$ref" : "#/definitions/JsonPrimitive"
        },
        "asLong" : {
          "format" : "int64",
          "type" : "integer"
        },
        "asNumber" : {
          "$ref" : "#/definitions/Number"
        },
        "asShort" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asString" : {
          "type" : "string"
        },
        "empty" : {
          "type" : "boolean"
        },
        "jsonArray" : {
          "type" : "boolean"
        },
        "jsonNull" : {
          "type" : "boolean"
        },
        "jsonObject" : {
          "type" : "boolean"
        },
        "jsonPrimitive" : {
          "type" : "boolean"
        }
      }
    },
    "JsonPrimitive" : {
      "type" : "object",
      "properties" : {
        "asBigDecimal" : {
          "type" : "number"
        },
        "asBigInteger" : {
          "type" : "integer"
        },
        "asBoolean" : {
          "type" : "boolean"
        },
        "asByte" : {
          "format" : "byte",
          "type" : "string"
        },
        "asCharacter" : {
          "type" : "string"
        },
        "asDouble" : {
          "format" : "double",
          "type" : "number"
        },
        "asFloat" : {
          "format" : "float",
          "type" : "number"
        },
        "asInt" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asJsonArray" : {
          "$ref" : "#/definitions/JsonArray"
        },
        "asJsonNull" : {
          "$ref" : "#/definitions/JsonNull"
        },
        "asJsonObject" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "asJsonPrimitive" : {
          "$ref" : "#/definitions/JsonPrimitive"
        },
        "asLong" : {
          "format" : "int64",
          "type" : "integer"
        },
        "asNumber" : {
          "$ref" : "#/definitions/Number"
        },
        "asShort" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asString" : {
          "type" : "string"
        },
        "boolean" : {
          "type" : "boolean"
        },
        "jsonArray" : {
          "type" : "boolean"
        },
        "jsonNull" : {
          "type" : "boolean"
        },
        "jsonObject" : {
          "type" : "boolean"
        },
        "jsonPrimitive" : {
          "type" : "boolean"
        },
        "number" : {
          "type" : "boolean"
        },
        "string" : {
          "type" : "boolean"
        }
      }
    },
    "Number" : {
      "type" : "object"
    },
...
  },
  "host" : "localhost:8080",
  "info" : {
    "description" : "This is a sample.",
    "title" : "EJournal",
    "version" : "v1"
  },
  "paths" : {
...
    "/auth/setCodePep" : {
      "patch" : {
        "operationId" : "setCodePep",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataAuth"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
...
    "/parents/setCodePep" : {
      "post" : {
        "operationId" : "setCodePep",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataParents"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
...
    "/teachers/addTea" : {
      "post" : {
        "operationId" : "addTea",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataTeachers"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
    "/teachers/chPep" : {
      "post" : {
        "operationId" : "chPep",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataTeachers"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
    "/teachers/getTeachers" : {
      "post" : {
        "operationId" : "getTeachers",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataTeachers"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
    "/teachers/remPep" : {
      "post" : {
        "operationId" : "remPep",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataTeachers"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
    "/teachers/setCodePep" : {
      "post" : {
        "operationId" : "setCodePep",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataTeachers"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
...
  },
  "schemes" : [ "http" ],
  "swagger" : "2.0",
...
}